+++ src/app/cart/cart.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { SharedModule } from './../shared/shared.module';
+import { CartRoutingModule } from './cart-routing.module';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 
@@ -7,7 +8,8 @@ import { CartItemComponent } from './components/cart-item/cart-item.component';
 @NgModule({
   declarations: [CartListComponent, CartItemComponent],
   imports: [
-    SharedModule
+    SharedModule,
+    CartRoutingModule
   ],
   exports: [
     CartListComponent