+++ src/app/products/components/product-list/product-list.component.ts
@@ -3,6 +3,7 @@ import { Product } from './../../models/product.model';
 import { ProductsService } from './../../services/products.service';
 import { CommunicatorService } from './../../../shared/services/communicator.service';
 import { Observable, Subscriber } from 'rxjs';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-product-list',
@@ -13,7 +14,7 @@ export class ProductListComponent implements OnInit {
   products: Observable<Array<Product>>;
   products2: Promise<Array<Product>>;
   productsInBasket: Array<Product> = new Array<Product>();
-  constructor(private productService: ProductsService, private communicatorService: CommunicatorService) { }
+  constructor(private router: Router, private productService: ProductsService, private communicatorService: CommunicatorService) { }
 
   ngOnInit(): void {
     this.products = this.productService.getProducts();
@@ -28,6 +29,12 @@ export class ProductListComponent implements OnInit {
     this.communicatorService.publishProduct(product);
 
     console.log('products count in basket ', this.productsInBasket.length);
+  }
+
+  onEditProduct(p: Product): void {
+    const link = ['/edit', p.id];
+    console.log('router navigat /edit', p.id);
+    this.router.navigate(link);
   }
 
 }