+++ src/app/products/models/product.model.ts
@@ -1,3 +1,8 @@
 export class Product {
-  constructor(public name: string, public description: string, public price: number, public isAvailable: boolean) {}
+  constructor(
+    public id: number = null,
+    public name: string,
+    public description: string,
+    public price: number,
+    public isAvailable: boolean) {}
 }